move objnode handling to common code
objnode is used OS agnostic and used only
by dmu_redact.c
